ai.py
```python
import random
from config import Config

class AI:
    def __init__(self):
        self.config = Config()

    def generate_environment(self, game_env):
        """
        Generate a dynamic game environment using AI.
        """
        # Generate a random environment based on the configuration
        environment = self._generate_random_environment()

        # Set the generated environment to the game environment
        game_env.environment = environment

        return game_env

    def _generate_random_environment(self):
        """
        Generate a random environment based on the configuration.
        """
        environment = []

        for _ in range(self.config.environment_size):
            # Generate a random terrain type based on the configuration
            terrain_type = random.choice(self.config.terrain_types)

            # Generate a random enemy type based on the configuration
            enemy_type = random.choice(self.config.enemy_types)

            # Generate a random item type based on the configuration
            item_type = random.choice(self.config.item_types)

            # Add the generated terrain, enemy, and item to the environment
            environment.append({
                'terrain': terrain_type,
                'enemy': enemy_type,
                'item': item_type,
            })

        return environment
```
